refactor(frontend): Split evaluation modal and snippet list
