+++ frontend/src/app/assignment/modules/solution/edit-snippet/edit-snippet.component.ts
@@ -2,7 +2,7 @@ import {Component, EventEmitter, Input, Output} from '@angular/core';
 import {merge, OperatorFunction, Subject} from 'rxjs';
 import {debounceTime, distinctUntilChanged, map} from 'rxjs/operators';
 import {Snippet} from '../../../model/evaluation';
-import {selectionComment} from '../evaluation-modal/evaluation-modal.component';
+import {selectionComment} from '../snippet-list/snippet-list.component';
 
 @Component({
   selector: 'app-edit-snippet',
