+++ frontend/src/app/assignment/modules/solution/snippet-list/snippet-list.component.html
@@ -1,0 +1,50 @@
+<label class="form-label" for="snippets">
+  Snippets
+</label>
+<div *ngIf="searchSummary" class="alert alert-{{searchSummary.level}}">
+  Code Search found
+  <b>{{ searchSummary.hits }}</b>
+  hits in
+  <b>{{ searchSummary.files }}</b>
+  files from
+  <b>{{ searchSummary.solutions }}</b>
+  solutions for this snippet.
+  {{ searchSummary.message }}
+  <a [routerLink]="['/assignments', route.snapshot.params.aid, 'search']"
+     [queryParams]="{q: searchSummary.code, glob: task?.glob}">
+    View Results
+  </a>
+</div>
+<ul class="list-group" id="snippets">
+  <li *ngFor="let snippet of dto.snippets; let index=index" class="list-group-item"
+      [class.list-group-item-warning]="snippet.comment === selectionComment">
+    <app-edit-snippet
+      [snippet]="snippet" [index]="index" [comments]="comments"
+      (deleted)="deleteSnippet(index)"
+    ></app-edit-snippet>
+    <app-snippet [snippet]="snippet" wildcard="***" (updated)="snippetUpdates$.next($event)"></app-snippet>
+  </li>
+</ul>
+<div class="form-text">
+  Use
+  <a href="https://marketplace.visualstudio.com/items?itemName=fulib.fulibFeedback" target="_blank">
+    fulibFeedback
+  </a>
+  to add code snippets from within your IDE.
+  <span *ngIf="task && task.glob" class="text-info">
+        Code Search will only look for files matching the glob pattern <code>{{ task.glob }}</code>.
+      </span>
+</div>
+<ng-container *ngIf="snippetSuggestionsEnabled && embeddingSnippets.length">
+  <label class="form-label" for="embedding-snippets">
+    Snippet Suggestions
+    <span class="badge bg-gradient-primary">New</span>
+    &nbsp;
+    <i class="bi-question-circle" [ngbTooltip]='"Code Search suggests code that is related to the task \"" + task?.description + "\"" '></i>
+  </label>
+  <ul class="list-group" id="embedding-snippets">
+    <li *ngFor="let snippet of embeddingSnippets; let first=first" class="list-group-item">
+      <app-snippet [snippet]="snippet" [expanded]="first" (confirmed)="confirmEmbedding(snippet)"></app-snippet>
+    </li>
+  </ul>
+</ng-container>
